requirements.txt
acres==0.2.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
etelemetry==0.3.1
executing==2.2.0
feedparser==6.0.11
filelock==3.17.0
fitz==0.0.1.dev2
Flask==3.0.3
frontend==0.0.3
google-auth==2.38.0
google-genai==1.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
looseversion==1.3.0
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.3
openai==1.63.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20240706
pexpect==4.9.0
pikepdf==9.5.2
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.51
prov==2.0.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
PyGithub==2.6.0
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.3
PyNaCl==1.5.0
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
pyxnat==1.6.3
pyzmq==26.4.0
rdflib==6.3.2
requests==2.32.3
retrying==1.3.4
rsa==4.9
scipy==1.15.1
setuptools==75.8.2
sgmllib3k==1.0.0
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.45.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.1
uritools==4.0.3
urlextract==1.9.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.21.0
